refactor(smb): split set-info by info class, security, and lease breaks
